LayoutTests/fast/canvas/webgl/data-view-crash.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
description("Test that DataView does not crash with bad offset or length.");

var array = new Uint8Array([164, 112, 157, 63]);
var view;
shouldThrow("view = new DataView(array.buffer, -4500000000)");
shouldThrow("view = new DataView(array.buffer, -4500000000, 4500000000)");
var value = view ? view.getFloat32(0, true) : 0;
</script>

<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
